requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
praw==7.6.0
prawcore==2.3.0
python-dotenv==0.20.0
requests==2.28.0
update-checker==0.18.0
urllib3==1.26.9
websocket-client==1.3.2
py-cord==2.0.0b7